Refactor

+++ src/app/entities/pages/ArticlePage/index.ts
@@ -1,5 +1,6 @@
 import { IArticleService } from "../../../interfaces/services/ArticleService/types";
 import { INavigationService } from "../../../interfaces/services/NavigationService/types";
+import { IUserService } from "../../../interfaces/services/UserService/types";
 import { IArticle } from "../../components/Article/types";
 import { Control } from "../../components/Control";
 import { IControl } from "../../components/Control/types";
@@ -7,6 +8,8 @@ import { Field } from "../../components/Field";
 import { getNavigationTabs } from "../../components/Navigation";
 import { ExclusiveSelector } from "../../components/Selector/ExclusiveSelector";
 import { TTab } from "../../components/Tab/types";
+import { IUser } from "../../components/User/types";
+import { FollowControl } from "../../components/FollowControl";
 import { EPage, IPage } from "../types";
 
 export class ArticlePage implements IPage {
@@ -17,6 +20,7 @@ export class ArticlePage implements IPage {
 
   public comment = new Field<string>("");
 
+  public followControl: IControl | undefined;
   public editControl: IControl | undefined;
   public deleteControl: IControl | undefined;
   public submitCommentControl: IControl | undefined;
@@ -24,6 +28,7 @@ export class ArticlePage implements IPage {
   protected constructor(
     protected articleService: IArticleService,
     protected navigationService: INavigationService,
+    protected userService: IUserService,
   ) {
     if (!articleService) throw new Error("No article service provided");
     if (!navigationService) throw new Error("No navigation service provided");
@@ -35,8 +40,13 @@ export class ArticlePage implements IPage {
     articleId: string,
     articleService: IArticleService,
     navigationService: INavigationService,
+    userService: IUserService,
   ) {
-    const page = new ArticlePage(articleService, navigationService);
+    const page = new ArticlePage(
+      articleService,
+      navigationService,
+      userService,
+    );
 
     await page.initialize(articleId);
 
@@ -45,24 +55,32 @@ export class ArticlePage implements IPage {
 
   public async initialize(articleId: string): Promise<void> {
     this.article = await this.articleService.prepareArticle(articleId);
+    const username = this.article?.articleData.username;
+    const user: IUser | undefined =
+      username && (await this.userService.getUserProfile(username));
 
-    this.editControl = new Control("Change", async () => {
-      const id = this.article?.articleData.id;
-      if (!id) return;
-      await this.navigationService.navigateToArticle(id, true);
-    });
+    if (username === this.userService.currentUser) {
+      this.editControl = new Control("Change", async () => {
+        const id = this.article?.articleData.id;
+        if (!id) return;
+        await this.navigationService.navigateToArticle(id, true);
+      });
 
-    this.deleteControl = new Control("Delete", async () => {
-      const id = this.article?.articleData.id;
-      if (!id) return;
-      await this.articleService.delete(id);
-    });
+      this.deleteControl = new Control("Delete", async () => {
+        const id = this.article?.articleData.id;
+        if (!id) return;
+        await this.articleService.delete(id);
+      });
+    }
 
     this.submitCommentControl = new Control("Submit", async () => {
       const id = this.article?.articleData.id;
       if (!id) return;
       await this.articleService.submitComment(this.comment.value, id);
+      this.article = await this.articleService.prepareArticle(articleId);
     });
+
+    this.followControl = user && new FollowControl(user);
 
     this.submitCommentControl.isDisabled = true;
     this.comment = new Field<string>("", async () => {
